Crate cranelift_native

Performs autodetection of the host for the purposes of running Cranelift to generate code to run on the same machine.

Constants

VERSION
Version number of this crate.

Functions

builder
Return an isa builder configured for the current host machine, or Err(()) if the host machine is not supported in the current configuration.